refactor: common space
