+++ packages/web/src/features/common-space/services/getCommonSpaces.ts
