+++ packages/web/src/features/my/frames/MyServiceFrame.tsx
@@ -6,8 +6,8 @@ import AsyncBoundary from "@sparcs-clubs/web/common/components/AsyncBoundary";
 import FlexWrapper from "@sparcs-clubs/web/common/components/FlexWrapper";
 import FoldableSectionTitle from "@sparcs-clubs/web/common/components/FoldableSectionTitle";
 import MoreDetailTitle from "@sparcs-clubs/web/common/components/MoreDetailTitle";
+import MyCommonSpaceTable from "@sparcs-clubs/web/features/common-space/components/MyCommonSpaceTable";
 import MyActivityCertificateTable from "@sparcs-clubs/web/features/my/components/MyActivityCertificateTable";
-import MyCommonSpaceTable from "@sparcs-clubs/web/features/my/components/MyCommonSpaceTable";
 import MyPrintingTable from "@sparcs-clubs/web/features/my/components/MyPrintingTable";
 import MyRentalTable from "@sparcs-clubs/web/features/rental-business/components/MyRentalTable";
 import { useGetMyRentals } from "@sparcs-clubs/web/features/rental-business/services/useGetMyRentals";
